requirements.txt
appdirs==1.4.4
dataclasses==0.7
decorator==4.4.2
future==0.18.2
joblib==0.17.0
Mako==1.1.3
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.19.2
Pillow==8.0.1
polygraphy==0.20.13
pretrainedmodels==0.7.4
ptflops==0.6.3
pycuda==2020.1
pytools==2020.4.3
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
tensorrt==7.2.1.6
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
torch==1.7.0
torchvision==0.8.1
tqdm==4.54.0
typing-extensions==3.7.4.3
onnxruntime-gpu==1.6.0
pandas==1.1.3
plotly==4.14.1